pyproject.toml
[project]
name = "integracar-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "django>=5.2.6",
    "django-cotton>=2.1.3",
    "django-htmx>=1.26.0",
    "django-q2>=1.8.0",
    "django-widget-tweaks>=1.5.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.8",
    "langchain-ollama>=0.3.8",
    "langchain-openai>=0.3.33",
    "langchain-voyageai>=0.1.7",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-ai>=1.0.10",
    "pydantic-settings>=2.10.1",
    "pypdf>=6.1.0",
    "taskipy>=1.14.1",
]


[dependency-groups]
dev = [
    "djlint>=1.36.4",
    "ruff>=0.13.1",
]


[tool.taskipy.tasks]
# Django
run = "python manage.py runserver"
makemigrations = "python manage.py makemigrations"
migrate = "python manage.py migrate"
collectstatic = "python manage.py collectstatic --noinput -c"
createsuperuser = "python manage.py createsuperuser"

# outros
lint = "ruff check . && djlint . --quiet"
format = "ruff check . --fix && ruff format . && djlint . --reformat --quiet"

[tool.ruff]
line-length = 79
extend-exclude = ["migrations"]


[tool.ruff.lint]
preview = true
select = ["A", "B", "C", "E", "F", "I", "PL", "PT", "Q", "S", "T", "W", "DJ"]
ignore = ["E501", "PLR6301"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.format]
preview = true
quote-style = "single"

[tool.djlint]
max_blank_lines = 1
indent = 4
tabs = true
format_css = true
format_js = true
use_gitignore = true
space_around_blocks = true
blank_line_before_tag = "include,block,endblock,crispy,if,else,endif,for,endfor,comment,endcomment,with,endwith"
blank_line_after_tag = "include,block,endblock,crispy,if,else,endif,for,endfor,comment,endcomment,with,endwith,load,extends"
ignore = "H005,H006,H007,H008,H009,H010,H011,H012,H013,H014,H015,H016,H017,H018,H019,H020,H021,H022,H023,H024,H025,H026,H027,H028,H029,H030,H031,H033,H035,H036,H037,J004,J018,T001,T002,T003,T027,T028,T032,T034"
custom_html = "c-"

[tool.djlint.css]
indent_size = 4

[tool.djlint.js]
indent_size = 4
